adjust.flowPeaks {flowPeaks}	R Documentation

Adjusting the smoothing and merging behavior of the flowPeaks results

Description

Adjusting the smoothing and merging behavior of the flowPeaks results by changing the multiplers of the covariance matrix and the tolerance level for joining two peaks

Usage

adjust.flowPeaks(object,tol,h0,h,...)

Arguments

object
The output from the function flowPeaks

tol
See flowPeaks

h0
See flowPeaks

h
See flowPeaks

...
Optional additional arguments. At present no additional arguments are used.

Value

It returns an updated object of class flowPeaks, the detail defintion of which can be seen in flowPeaks.

Author(s)

Yongchao Ge yongchao.ge@gmail.com

See Also

flowPeaks

[Package flowPeaks version 0.99.6 Index]