evalCluster {flowPeaks}	R Documentation

evaulate the result of a clustering algorihm by comparing it with the gold standard

Description

This function takes the cluster labels of the two clusterings, one is based on the gold standard, the other is a candidate clusterign, and compute one of the three metrics to assess the candidate clustering performance.

Usage

evalCluster(gs,cand,method=c("Rand.index","Fmeasure","Vmeasure"),
                      rm.gs.outliers=TRUE)

Arguments

gs
A integer-valued vector of length n for the cluster labels of the gold standard clustering, where negative numbers such as -1 is for the outerliers

cand
A integer-valued vector of length n for the cluster label of a candidate clustering, where -1 is for the outliers

rm.gs.outliers
Determining whether the outliers of the gold standard clustering should be removed in the comparison

method
A single character to indicate which one of three metrics should be used to evaluate the clustering. The details are described in Ge (2012) and references mentioned in that paper

Rand.index

The adjusted Rand.index

Fmeasure

F-measure

Vmeasure

V-measure

Author(s)

Yongchao Ge yongchao.ge@gmail.com

References

Ge Y. et al, flowPeaks: a fast unsupervised clustering for flow cytometry data via K-means and density peak finding, 2012, Bioinformatics, in press.

See Also

flowPeaks

[Package flowPeaks version 0.99.6 Index]