plot.flowPeaks {flowPeaks}	R Documentation

Plot the results generated by flowPeaks

Description

This function takes the results generated from flowPeaks as an input, and plot the data in 2D. These plots display the clustering structure

Usage

## S3 method for class 'flowPeaks'
plot(x,idx=c(1,2),drawlab=FALSE,
cols=c("red","green3","blue","cyan","magenta","yellow","gray"),drawvor=TRUE,
                         drawlocalpeaks=FALSE,drawkmeans=TRUE,drawboundary=TRUE,
                         classlab, negcol, negpch,...)

Arguments

x
Anobject of class flowPeaks, e.g., t the output from the functions flowPeaks or adjust.flowPeaks

idx
The index of the columns will be used to plot the clustering. idx needs to be at least legnth 2, and have no duplicate elements, and the values can only take from 1 to d, where d is the number of columns for the input matrix x that is used as an input of the function flowPeaks

drawlab
The option to decide whether we should draw the cluster labels

cols
The color specification for plotting the points in each cluster. Please note, "white" and "black" are not allowed, which are reserved for other purpse

drawvor
Deciding whether the voronoi diagram should be drawn, only good for 2D data

drawlocalpeaks
Decding whether the local peaks with a triangle symbol should be drawn

drawkmeans
Deciding whether the kmeans center with a filled circle should be drawn

drawboundary
Deciding whether the boudary between clusters should be drawn, only good for 2D data

classlab
Use this to replace the default class labels from x$peak.cluster, for example, the classlab may come from assign.flowPeaks

negcol
Deciding the color of the negative, which are outliers

negpch
Deciing the symbols for the outliers

...
Optional additional arguments. At present no additional arguments are used.

Author(s)

Yongchao Ge yongchao.ge@gmail.com

See Also

flowPeaks

[Package flowPeaks version 0.99.6 Index]